Update getHealth to be double

+++ src/main/java/org/dynmap/common/DynmapPlayer.java
@@ -44,7 +44,7 @@ public interface DynmapPlayer extends DynmapCommandSender {
     /**
      * Get health
      */
-    public int getHealth();
+    public double getHealth();
     /**
      * Get armor points
      */
